requirements.txt
openai
Flask
Flask-CORS
python-dotenv 